// src/api/generated/models/paymentSummaryStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type PaymentSummaryStatus = typeof PaymentSummaryStatus[keyof typeof PaymentSummaryStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PaymentSummaryStatus = {
  PENDING: 'PENDING',
  UPLOADED: 'UPLOADED',
  CONFIRMED: 'CONFIRMED',
  REJECTED: 'REJECTED',
} as const;